phonebusted/requirements.txt
ultralytics
opencv-python
pystray
Pillow
requests
pyttsx3
